1) Refactored Administrator and Consultant entities (added AbstractUserEntity). 2) Improved exception handler. 3) Separated Administrator's and Consultant's JWT validations and accept. 4) Merged all functions of few JWT generator classes to one class.

+++ src/main/java/carDealerApplication/api/controller/AdministratorController.java
@@ -28,49 +28,49 @@ public class AdministratorController {
 
     @GetMapping("car")
     public List<Car> fetchCarList(@RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         return carService.fetchEntityList();
     }
 
     @GetMapping("dealer_center")
     public List<DealerCenter> fetchDealerCenterList(@RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         return dealerCenterService.fetchEntityList();
     }
 
     @GetMapping("manufacturer")
     public List<Manufacturer> fetchManufacturerList(@RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         return manufacturerService.fetchEntityList();
     }
 
     @GetMapping("consultant")
     public List<ConsultantDTO> fetchConsultantList(@RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         return consultantService.fetchEntityList();
     }
 
     @GetMapping("administrator")
     public List<AdministratorDTO> fetchAdministratorList(@RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         return administratorService.fetchEntityList();
     }
 
     @GetMapping("special_offer")
     public List<SpecialOffer> fetchSpecialOfferList(@RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         return specialOfferService.fetchEntityList();
     }
 
     @GetMapping("location")
     public List<Location> fetchLocationList(@RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         return locationService.fetchEntityList();
     }
 
     @PutMapping("car")
     public void upsertCar(@RequestBody Car car, @RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         administratorApiValidator.validateCar(car);
         carService.upsertEntity(car);
     }
@@ -78,7 +78,7 @@ public class AdministratorController {
     @PutMapping("dealer_center")
     public void upsertDealerCenter(@RequestBody DealerCenter dealerCenter,
                                    @RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         administratorApiValidator.validateDealerCenter(dealerCenter);
         dealerCenterService.upsertEntity(dealerCenter);
     }
@@ -86,7 +86,7 @@ public class AdministratorController {
     @PutMapping("manufacturer")
     public void upsertManufacturer(@RequestBody Manufacturer manufacturer,
                                    @RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         administratorApiValidator.validateManufacturer(manufacturer);
         manufacturerService.upsertEntity(manufacturer);
     }
@@ -94,7 +94,7 @@ public class AdministratorController {
     @PutMapping("consultant")
     public void upsertConsultant(@RequestBody ConsultantDTO consultantDTO,
                                  @RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         administratorApiValidator.validateConsultant(consultantDTO);
         consultantService.upsertEntity(consultantDTO);
     }
@@ -102,7 +102,7 @@ public class AdministratorController {
     @PutMapping("administrator")
     public void upsertAdministrator(@RequestBody AdministratorDTO administratorDTO,
                                     @RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         administratorApiValidator.validateAdministrator(administratorDTO);
         administratorService.upsertEntity(administratorDTO);
     }
@@ -110,14 +110,14 @@ public class AdministratorController {
     @PutMapping("special_offer")
     public void upsertSpecialOffer(@RequestBody SpecialOffer specialOffer,
                                    @RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         administratorApiValidator.validateSpecialOffer(specialOffer);
         specialOfferService.upsertEntity(specialOffer);
     }
 
     @DeleteMapping("car/delete/{id}")
     public String deleteCar(@PathVariable("id") Long id, @RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         carService.deleteEntityById(id);
         return "This car was successfully deleted for good";
     }
@@ -125,7 +125,7 @@ public class AdministratorController {
     @DeleteMapping("dealer_center/delete/{id}")
     public String deleteDealerCenter(@PathVariable("id") Long id,
                                      @RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         dealerCenterService.deleteEntityById(id);
         return "This dealer center was successfully deleted for good";
     }
@@ -133,7 +133,7 @@ public class AdministratorController {
     @DeleteMapping("manufacturer/delete/{id}")
     public String deleteManufacturer(@PathVariable("id") Long id,
                                      @RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         manufacturerService.deleteEntityById(id);
         return "This manufacturer was successfully deleted for good";
     }
@@ -141,7 +141,7 @@ public class AdministratorController {
     @DeleteMapping("consultant/delete/{id}")
     public String deleteConsultant(@PathVariable("id") Long id,
                                    @RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         consultantService.deleteEntityById(id);
         return "This consultant was successfully erased for good";
     }
@@ -149,7 +149,7 @@ public class AdministratorController {
     @DeleteMapping("administrator/delete/{id}")
     public String deleteAdministrator(@PathVariable("id") Long id,
                                       @RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         administratorService.deleteEntityById(id);
         return "This administrator was successfully erased for good";
     }
@@ -157,7 +157,7 @@ public class AdministratorController {
     @DeleteMapping("special_offer/delete/{id}")
     public String deleteSpecialOffer(@PathVariable("id") Long id,
                                      @RequestHeader(HttpHeaders.AUTHORIZATION) String jwtToken) {
-        jwtHandlerUtil.validateJwt(jwtToken);
+        jwtHandlerUtil.validateJwtForAdministrator(jwtToken);
         specialOfferService.deleteEntityById(id);
         return "This car was successfully deleted for good";
     }
